==> Python/01-STR16.ipynb <==
import pandas as pd
import numpy as np

pd.set_option('display.max_rows', 200)
df = pd.read_csv('ready_data00.csv', low_memory=False)
df.head(1)

df.info()

df.transaction_type = df.transaction_type + ' ' + df.additional_transaction_info.fillna('')
df.transaction_type = df.transaction_type.str.strip()


df.commodity = df.commodity.astype('category')
df.transaction_type = df.transaction_type.astype('category')
df.country_or_area = df.country_or_area.astype('category')
df.commodity_transaction = df.commodity_transaction.astype('category')
df.unit = df.unit.astype('category')
df.category = df.category.astype('category')
df.year = df.year.astype('category')
df.quantity = df.quantity.astype('float32')
df.quantity_footnotes = df.quantity_footnotes.astype('float32')
df.info()

only_comm = df.loc[:,['country_or_area','year',"transaction_type",'unit', 'quantity', 'category']]
only_comm.head()

only_comm["transaction_type"].value_counts().sort_index()

#CL1214a - CL1214o -> CL1214

other_manuf_const_and_non_fuel_min_ind = ['consumption by non-ferrous metals',
                                          'consumption by non-metallic minerals',
                                          'consumption by transport equipment',
                                          'consumption by machinery',
                                          'consumption by mining and quarrying',
                                          'consumption by food and tobacco',
                                          'consumption by food, beverage and tobacco',
                                          'consumption by paper, pulp and print',
                                          'consumption by wood and wood products',
                                          'consumption by construction',
                                          'consumption by industry & construction',
                                          'consumption by textile and leather', 
                                          'consumption not elsewhere specified (industry)',
                                          'consumption by industries not elsewhere specified'
                                           ]

def CL4(row):
    if row in other_manuf_const_and_non_fuel_min_ind:
        return 'other manuf., const. and non-fuel min. ind.'
    else:
        return np.nan
        
only_comm['CL4'] =  only_comm['transaction_type'].apply(CL4)

only_comm.head()

#CL08811 - CL08832 - > CL088

transformation_in_electricity_chp_and_heat_plants = ['main activity',
                                                     'main activity electricity plants',
                                                     'main activity chp plants ',
                                                     'main activity total',
                                                     'main activity heat plants',
                                                     'transformation in main activity producer electricity plants',
                                                     'transformation in main activity producer chp plants',
                                                     'transformation in main activity producer heat plants',
                                                     'transformation in main activity electricity plants',
                                                     'transformation in main activity heat plants',
                                                     'transformation in main activity chp plants',
                                                     'transformation in chp plants main activity producers',
                                                     'transformation in electricity plants main activity producers',
                                                     'transformation in heat plants main activity producers',
                                                     'autoproducer',
                                                     'autoproducer electricity plants',
                                                     'autoproducer chp plants',
                                                     'autoproducer heat plants',
                                                     'autoproducer total',
                                                     'transformation in autoproducer electricity plants',                                 
                                                     'transformation in autoproducer chp plants',
                                                     'transformation in autoproducer heat plants',
                                                     'transformation in chp plants autoproducers',
                                                     'transformation in electricity plants autoproducers',
                                                     'transformation in heat plants autoproducers',
                                                     'other transformation processes not elsewhere classified',
                                                     'from other sources main activity producers',
                                                     'from other sources autoproducers',
                                                     'from other sources total'
                                                    ]

#CL1211 - CL1214 - > CL121
manufacturing_construction_and_non_fuel= ['consumption by iron and steel',
                                          'consumption by chemical and petrochemical',
                                          'consumption by chemical',
                                          'consumption by other manuf., const. and non-fuel',
                                          'consumption by other manufacturing, construction and non-fuel'
                                            ]


#CL1222-CL1225 -> CL122
transport = ['consumption by rail',
             'consumption by road',
             'consumption by domestic navigation',
             'consumption by domestic aviation',
             'consumption by pipeline transport',
             'consumption by other transport',
             'consumption not elsewhere specified (transport)'
            ]


#CL1231-1235 -> CL123
other = ['consumption by households',
        'consumption by agriculture, forestry and fishing',
        'consumption by commercial and public services',
        'consumption by other consumers',
        'consumption by other consumers not elsewhere specified',
        'consumption by other industries and construction',
        'consumption by other',
        'commercial and public services',
        'commercial and public services',
        'consumption not elsewhere specified (other)'
        ]

#cl071
net = [ "net installed capacity of electric power plants public solar",
        "net installed capacity of electric power plants, public combustible fuels",
        "net installed capacity of electric power plants, public geothermal", 
        "net installed capacity of electric power plants, public hydro",  
        "net installed capacity of electric power plants, public nuclear",
        "net installed capacity of electric power plants, public tide, wave, marine",
        "net installed capacity of electric power plants, public wind",
        "net installed capacity of electric power plants, self-producer combustible fuels", 
        "net installed capacity of electric power plants, self-producer geothermal",
        "net installed capacity of electric power plants, self-producer hydro",
        "net installed capacity of electric power plants, self-producer nuclear",  
        "net installed capacity of electric power plants, self-producer solar",
        "net installed capacity of electric power plants, self-producer tide, wave, marine",
        "net installed capacity of electric power plants, self-producer wind", 
        "net production",
        "net transfers",
        "pumped hydro main activity producers",
        "solar pv main activity producers",
        "solar thermal main activity producers",
        "pumped hydro autoproducers",
        'solar pv autoproducers',
        'solar thermal autoproducers',
        'pumped hydro total',
        'solar pv total',
        'solar thermal total'
          ]



def CL3(row):
    if row['transaction_type'] in transformation_in_electricity_chp_and_heat_plants:
        return 'transformation in electricity, chp and heat plants'
    elif row['transaction_type'] in net:
        return 'Transfers details'
    elif row['transaction_type'] in manufacturing_construction_and_non_fuel:
        return 'consumption by manufacturing, construction and non-fuel'
    elif row['transaction_type'] in transport:
        return 'transport'
    elif row['transaction_type'] in other :
        return 'other'
    else:
        return np.nan
    
    
only_comm['CL3'] =  only_comm.apply(CL3, axis=1)

production = ['production from refineries', 
              'production from plants' ,
              'production from off-shore',
              'total production, main activity',
              'total nuclear production', 
              'used for electricity production'
             ]

international_marine_bunkers = ['international marine bunkers',
                                'international aviation bunkers']
#CL088 - CL089 -> CL08
transformation = [# usunięcie 'transformation in electricity, chp and heat plants', 
                    'transformation in coke ovens',
                    'transformation in gas works',
                    'transformation in briquetting plants',
                    'transformation in blast furnaces',
                    'transformation in coal liquefaction plants',
                    'transformation in oil refineries',
                    'transformation in charcoal plants',
                    'transformation in petrochemical plants',
                    'transformation in natural gas blending plants',
                    'transformation in gas-to-liquid plants',
                    'used for electric boilers',
                    'used for heat pumps',
                    'other transformation'
                 ]

#CL0911-CL0928 -> CL09
energy_industries_own_use = ['own use by coal mines',
                            'own use by coke ovens',
                            'own use by gas works',
                            'own use by briquetting plants',
                            'own use by blast furnaces',
                            'own use by oil refineries',
                            'own use by oil and gas extraction',
                            'own use by electricity, chp and heat plants',
                            'own use by coal liquefaction plants',
                            'own use by pump-storage plants',
                            'own use by lng/regasification plants',
                            'own use by nuclear fuel extraction and processing',
                            'own use by biogas production plants',
                            'own use by other energy producers',
                            'own use by gas-to-liquid (gtl) plants',
                            'own use by charcoal plants',
                            'own use by gasworks',
                            'consumption by lng/regasification plants',
                            'consumption by coal mines',
                            'consumption by energy sector',
                            'consumption by briquetting plants', 
                            'conversion in briquetting plants',
                            'conversion in blast furnaces',
                            'other energy industry own use'
                            ]


#CL121-CL123 ->CL12
final_energy_consumption = ['consumption by manufacturing, construction and non-fuel',
                            'consumption by other manuf., const. and non-fuel',
                            'consumption by transport',
                            'consumption by electricity, chp and heat plants'
                            'transport', 
                            'other'
                           ]



#CL07
transfers = [   'total net installed capacity of electric power plants, main activity',
                'total net installed capacity of electric power plants, main activity & autoproducer',
                'total net installed capacity of electric power plants, combustible fuels',
                'total net installed capacity of electric power plants, hydro',
                'total net installed capacity of electric power plants, autoproducer',
                'total net installed capacity of electric power plants, wind',
                'total net installed capacity of electric power plants, solar',
                'total net installed capacity of electric power plants, nuclear',
                'total net installed capacity of electric power plants, geothermal',
                'total net installed capacity of electric power plants, tide, wave, marine',
                'total wind production',
                'total thermal production',
                'total hydro production',
                'total hydraulic resources',
                'total production, autoproducer',
                'total solar production',
                'total geothermal production',
                'total tide, wave production',
                'transfers and recycled products'
              ]  


losses = [ 'extraction loss shrinkage', 'flared' ]

def CL2(row):
    if row['transaction_type'] in production :
        return 'production'
    elif row['transaction_type'] == 'receipts from other sources':
        return ' receipts from other sources'
    elif row['transaction_type'] == 'imports': 
        return 'imports'
    elif row['transaction_type'] == 'exports': 
        return 'exports'
    elif row['transaction_type'] in international_marine_bunkers:
        return 'international marine bunkers'
    elif row['transaction_type'] == 'stock changes':
        return 'stock changes'
    elif row['transaction_type'] in net:
         return 'transfers'
    elif row['transaction_type'] in transformation:
        return 'transformation'
    elif row['transaction_type'] in energy_industries_own_use:
        return 'energy industries own use'
    elif row['transaction_type'] in losses:
        return 'losses'
    elif row['transaction_type'] == 'consumption for non-energy use':
        return 'non-energy use'
    elif row['transaction_type'] in final_energy_consumption:
        return 'final energy consumption'
    else:
        return np.nan
    
only_comm['CL2'] =  only_comm.apply(CL2, axis=1)

only_comm.info()

only_comm[only_comm['CL2'].isnull()]['transaction_type'].value_counts()

##nadgrupy
main_category = ['total energy supply','final consumption' ]


#nie wiem gdzie to zaliczyc
wtf = ['refinery capacity', 'gross demand','gross production','reserves' , 
       'memo:gross production' ,'flared and vented ' ,'total resources',
       'total resources in place','re-injected', 'total known reserves',
       'recoverable reserves', 'vented', 'additional resources', 'reserves from oil shale and tar sands',
       'reserves from oil shale', 'reasonably assured reserves', 'recoverable','reserves from tar sands',
       'estimated additional reserves', 'other reserves', 'memo: opening stock level (national territory)', 'memo: closing stock level (national territory)' 
      ]

only_comm2 = only_comm

only_comm2.insert(9,'commodity_transaction', df.commodity_transaction)
only_comm2.insert(10,'commodity', df.commodity)
only_comm2.to_csv('transaction_type013.csv', index = False)

only_comm2
