==> Python/04_electricity.ipynb <==
### Ulepszona wersja 040_electricity

import pandas as pd
import numpy as np

df = pd.read_csv('ready_data_030.csv',low_memory=False)

df.info()

## INFO
Kategoria Gross skada się sumy:
<ul>
    <li>thermal</li><li>nuclear</li><li>geothermal</li><li>hydro</li><li>solar</li><li>tide</li><li>wind</li><li>dodatkowo z: chemical heat – Autoproducer, chemical heat – Main activity, other sources – Autoproducer,From other sources – Main activity</li>


def get_country_electircity_stats(data,country,year):
    output_list = []
    data_list = ['Electricity - Gross production','Electricity - net production',
                 'Electricity - total production, main activity','Electricity - total production, autoproducer',
                 'Electricity - total thermal production','Electricity - total nuclear production',
                 'Electricity - total geothermal production','Electricity - total hydro production',
                 'Electricity - total solar production','Electricity - total tide, wave production',
                 'Electricity - total wind production',
                 'From chemical heat – Autoproducer','From chemical heat – Main activity',
                 'From other sources – Autoproducer','From other sources – Main activity']
    short_names = ['Gross', 'net','production-main activity','total-autoproducer',
                   'thermal','nuclear','geothermal','hydro','solar','tide','wind',
                   'chem-heat-auto','chem-heat-main','other-auto','other-main']
    x=0
    for i in data_list: 
        z = data['new_quantity'][(data['commodity_transaction'] == i) 
        & (data['country_or_area'] == country) & (data['year'] == year) & (data['category'] != 'heat')]
#         country = z
        if z.empty:
            z=pd.Series(0)
        z=pd.Series(z)
        output_list.append([short_names[x],z.iloc[0]])
        x+=1
    data_frame = pd.DataFrame(output_list)
    production_diff = data_frame.iloc[0,1].sum() - data_frame.iloc[4:15,1].sum()
    if_diff = 'difference beetwen gross production and sum of poduction in particular categories is: {}'.format(production_diff)

    return(output_list,data_frame,if_diff,country)

    


(a,b,c,d) = get_country_electircity_stats(df,'Finland',2006)
print(a)
print('\n')
print(c)
print(d)
b

# Test czy gross równa się sumie podkategorii, dla danego kraju w zakresie lat 1995-2015
for i in range(1995,2015):
    (a,b,c,d) = get_country_electircity_stats(df,'Finland',i)
    print(i,d,':',c)

#Test czy gross równa się sumie podkategorii, dla podanego zakresu lat we wszystkich krajach w bazie danych
countries = list(df.country_or_area.unique())
for i in range(2005,2010):
    for j in countries:
        (a,b,c,d) = get_country_electircity_stats(df,j,i)
        if d != 0:
            print(i,':', j,':',c)
        else:
            continue
            # licz = licz+1
            # print(licz)
