==> Python/05_do_analizy.ipynb <==
import pandas as pd


df = pd.read_csv('ready_data_030.csv', low_memory=False)
df


### wybrane kolumny

df = df[['country_or_area', 'ISO', 'year', 'new_unit', 'new_quantity','commodity_transaction', 'category']]
df

### wybrane kategorie

#List of columns selected for analysis
data_list = ['Electricity - Gross production',
             #'Electricity - total production, main activity', 
             #'Electricity - total production, autoproducer',
             'Electricity - total thermal production', 
             'Electricity - total nuclear production',
             'Electricity - total geothermal production', 
             'Electricity - total hydro production',
             'Electricity - total solar production',
             'Electricity - total tide, wave production',
             'Electricity - total wind production',
             'From chemical heat – Autoproducer',
             'From chemical heat – Main activity',
             'From other sources – Autoproducer',
             'From other sources – Main activity']
                 
df = df[df['commodity_transaction'].isin(data_list)]
df.head()

x = df[df['commodity_transaction'] == "Electricity - total hydro production"]['new_quantity'].sum()
y = df[df['commodity_transaction'] == 'Electricity - total tide, wave production']['new_quantity'].sum()

a = x+y
print('x: ', x, ' +  y: ',y, ' = ', a )

### Adding "Electricity - total hydro production" to 'Electricity - total tide, wave production'


dict_com_tr = {"Electricity - total hydro production":'Electricity - total tide, wave and hydro production',
         'Electricity - total tide, wave production': 'Electricity - total tide, wave and hydro production' }
dict_category = {'tide_wave_and_ocean_electricity': 'hydro_electricity',
                    'hydro':'hydro_electricity'}
df['commodity_transaction'] = df['commodity_transaction'].replace(dict_com_tr)
df['category'] = df['category'].replace(dict_category)
df.head()

df = df.groupby(['country_or_area','ISO','commodity_transaction','year','category','new_unit']).agg({'new_quantity':'sum'}).reset_index()
df.head(50)

### to check 

b = df[df['commodity_transaction'] == 'Electricity - total tide, wave and hydro production']['new_quantity'].sum()
print("b: ", b)

a-b

df.commodity_transaction.value_counts()

df.category.value_counts()

df['new_unit'] = 'TWh'
df['new_quantity'] = df['new_quantity']/1000000
df.drop(df[(df['commodity_transaction'] == 'From other sources – Main activity') & (df['category']=='heat')].index, inplace = True)
df.drop(df[(df['commodity_transaction'] == 'From other sources – Autoproducer') & (df['category']=='heat')].index, inplace = True)

## file creation

df.to_csv('final_version.csv', index= False)
df
